requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
requests>=2.31.0
boto3>=1.34.0
pytest>=8.0.0
httpx>=0.26.0
black>=24.1.0
ruff>=0.1.15
jinja2>=3.1.0
python-multipart>=0.0.6
celery>=5.3.0
redis>=5.0.0
PyJWT>=2.8.0
